requirements.txt
Flask==2.0.2
numpy==1.21.0
pandas==1.3.0
matplotlib==3.4.2
seaborn==0.11.1
scikit-learn==0.24.2
tensorflow==2.6.0
imbalanced-learn==0.8.1
Pillow==8.3.1
opencv-python==4.5.3.56
gunicorn==20.1.0
